upgrade oneapi to 2024.1 2/2 - pow(T1, T2) -> pow(T, T)

+++ tests/utils/profiling.hpp
@@ -103,7 +103,7 @@ class profiling_helper {
         //time mean square error
         for (int i = 1; i < iter; i++) {
 #if (__LIBSYCL_MAJOR_VERSION >= 7) && (__LIBSYCL_MINOR_VERSION >= 1)
-            stat.variance += sycl::pow(time[i] - stat.mean, 2);
+            stat.variance += sycl::pow(time[i] - stat.mean, (double)2);
 #else
             stat.variance += pow(time[i] - stat.mean, 2);
 #endif
